pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Вычислитель отличий"
authors = ["VictorVangeli <whgaleon1@yandex.ru>"]
readme = "README.md"
packages = [
  { include = "gendiff" },
]


[tool.poetry.dependencies]
python = "^3.10"
pathlib = "^1.0.1"
pyaml = "^24.9.0"


[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
black = "^24.8.0"
isort = "^5.13.2"
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
asciinema = "^2.4.0"


[tool.pytest.ini_options]
testpaths = ["tests"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
